# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Накопительная система лояльности «Гофермарт»",
        "title": "Gophermart API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/user/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Получение списка загруженных номеров заказов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешная обработка запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Order"
                            }
                        }
                    },
                    "204": {
                        "description": "нет данных для ответа",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Order"
                            }
                        }
                    },
                    "401": {
                        "description": "пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Загрузка номера заказа",
                "parameters": [
                    {
                        "description": "номер заказа",
                        "name": "orderNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "номер заказа уже был загружен этим пользователем"
                    },
                    "202": {
                        "description": "новый номер заказа принят в обработку"
                    },
                    "400": {
                        "description": "неверный формат запроса"
                    },
                    "401": {
                        "description": "пользователь не аутентифицирован"
                    },
                    "409": {
                        "description": "номер заказа уже был загружен другим пользователем"
                    },
                    "422": {
                        "description": "неверный формат номера заказа"
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "данные пользователя для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "пользователь успешно зарегистрирован и аутентифицирован",
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "токен авторизации"
                            }
                        }
                    },
                    "400": {
                        "description": "неверный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "409": {
                        "description": "логин уже занят",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Order": {
            "type": "object",
            "properties": {
                "accural": {
                    "type": "integer",
                    "example": 500
                },
                "number": {
                    "type": "string",
                    "example": "42"
                },
                "status": {
                    "description": "Статус расчёта начисления:\n* REGISTERED - заказ зарегистрирован, но вознаграждение не рассчитано;\n* INVALID - заказ не принят к расчёту, и вознаграждение не будет начислено;\n* PROCESSING - расчёт начисления в процессе;\n* PROCESSED - расчёт начисления окончен;",
                    "type": "string",
                    "enum": [
                        "REGISTERED",
                        "INVALID",
                        "PROCESSING",
                        "PROCESSED"
                    ],
                    "example": "PROCESSING"
                },
                "uploaded_at": {
                    "type": "string",
                    "example": "2025-06-23T23:48:45+03:00"
                }
            }
        },
        "api.UserRegisterRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}